% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{detach_NPSdataverse}
\alias{detach_NPSdataverse}
\title{Detach all loaded packages}
\usage{
detach_NPSdataverse()
}
\value{
invisilble
}
\description{
Detach all loaded packages
}
\examples{
\dontrun{
detach_NPSdataverse()
}
}
