% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{.latest_github_version}
\alias{.latest_github_version}
\title{Get the latest version of an R package that is on GitHub}
\usage{
.latest_github_version(repo, repo_owner, branch = "main", release_only = FALSE)
}
\arguments{
\item{repo}{Name of the GitHub repository (e.g. "DPchecker")}

\item{repo_owner}{Owner of the GitHub repository (e.g. "nationalparkservice")}

\item{branch}{Branch to use (defaults to "main")}

\item{release_only}{If TRUE, only looks at GitHub releases. If FALSE (default), looks at the version number in the DESCRIPTION file.}
}
\value{
A package version number
}
\description{
Get the latest version of an R package that is on GitHub
}
\keyword{internal}
