% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{.print_cust_package_deps}
\alias{.print_cust_package_deps}
\title{Custom print function for github repos to update}
\usage{
.print_cust_package_deps(x, show_ok = FALSE, ...)
}
\arguments{
\item{x}{dataframe. Output from NPSvers_update.}

\item{show_ok}{logical.}

\item{...}{}
}
\value{
printed text to console
}
\description{
formats print table for custom printing package dependencies in need to updating. \verb{Derived from remotes:::print.package_deps()}.
}
\keyword{internal}
