% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateR.R
\name{.update_git_repos}
\alias{.update_git_repos}
\title{Update info for git repos in NPSdataverse}
\usage{
.update_git_repos()
}
\value{
dataframe
}
\description{
\code{.update_git_repos()} determines which of the NPSdataverse packages from github need to be updated and provides instructions on how to do so.
}
\keyword{internal}
