% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{is_online}
\alias{is_online}
\title{Check internet connectivity}
\usage{
is_online(site = "https://github.com/")
}
\arguments{
\item{site}{string defaults to https://github.com}
}
\value{
logical
}
\description{
Checks whether the system can ping github.com. Adapted from:https://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r
}
\examples{
is_online()
}
