% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPSdataverse-package.R
\docType{package}
\name{NPSdataverse-package}
\alias{NPSdataverse}
\alias{NPSdataverse-package}
\title{NPSdataverse: Tools and Packages for Data and Metadata Manipulation}
\description{
Loads a suite of R packages for creating and manipulating data packages including interacting with DataStore.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nationalparkservice/NPSdataverse}
  \item Report bugs at \url{https://github.com/nationalparkservice/NPSdataverse/issues}
}

}
\author{
\strong{Maintainer}: Robert Baker \email{robert_baker@nps.gov} (\href{https://orcid.org/0000-0001-7591-5035}{ORCID})

Authors:
\itemize{
  \item Judd Patterson \email{Judd_Patterson@nps.gov} (\href{https://orcid.org/0000-0002-0951-7917}{ORCID})
  \item Joe DeVivo \email{Joe_DeVivo@nps.gov} (\href{https://orcid.org/0000-0002-0414-7865}{ORCID})
}

}
\keyword{internal}
