% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{NPSdataverse_detach}
\alias{NPSdataverse_detach}
\title{Detach all loaded packages}
\usage{
NPSdataverse_detach()
}
\value{
invisilble
}
\description{
Detach all loaded packages
}
\examples{
\dontrun{
NPSdataverse_detach()
}
}
